requirements.txt
pafy
youtube-dl
ffmpeg-python
wget
